Updated maint script to use fewer queries
Bug: T297484

+++ maintenance/MigrateSearchPagePrefs.php
@@ -49,6 +49,8 @@ class MigrateSearchPagePrefs extends Maintenance {
 
 		$this->addDescription( 'Migrates the sdms-specialsearch-default preferenc to '
 			. 'search-special-page ' );
+		$this->addOption( 'undo', 'Undo migration',
+			false, false, 'u' );
 	}
 
 	public function init() {
@@ -64,6 +66,14 @@ class MigrateSearchPagePrefs extends Maintenance {
 	public function execute() {
 		$this->init();
 
+		if ( $this->getOption( 'undo', false ) ) {
+			$this->undoMigration();
+		} else {
+			$this->migrate();
+		}
+	}
+
+	private function migrate() {
 		$userIds = $this->dbr->selectFieldValues(
 			'user_properties',
 			'up_user',
@@ -78,6 +88,13 @@ class MigrateSearchPagePrefs extends Maintenance {
 		}
 
 		$this->beginTransaction( $this->dbw, __METHOD__ );
+		$this->dbw->delete(
+			'user_properties',
+			[
+				'up_property' => [ 'sdms-specialsearch-default', 'search-special-page' ],
+			],
+			__METHOD__
+		);
 		$this->dbw->insert(
 			'user_properties',
 			array_map( static function ( $userId ) {
@@ -89,12 +106,40 @@ class MigrateSearchPagePrefs extends Maintenance {
 			}, $userIds ),
 			__METHOD__
 		);
+		$this->commitTransaction( $this->dbw, __METHOD__ );
+	}
+
+	private function undoMigration() {
+		$userIds = $this->dbr->selectFieldValues(
+			'user_properties',
+			'up_user',
+			[
+				'up_property' => 'search-special-page',
+				'up_value' => 'Search',
+			],
+			__METHOD__
+		);
+		if ( !$userIds ) {
+			return;
+		}
+
+		$this->beginTransaction( $this->dbw, __METHOD__ );
 		$this->dbw->delete(
 			'user_properties',
 			[
-				'up_user' => $userIds,
-				'up_property' => [ 'sdms-specialsearch-default' ],
+				'up_property' => [ 'search-special-page', 'sdms-specialsearch-default' ],
 			],
+			__METHOD__
+		);
+		$this->dbw->insert(
+			'user_properties',
+			array_map( static function ( $userId ) {
+				return [
+					'up_user' => $userId,
+					'up_property' => 'sdms-specialsearch-default',
+					'up_value' => '1',
+				];
+			}, $userIds ),
 			__METHOD__
 		);
 		$this->commitTransaction( $this->dbw, __METHOD__ );
